Move ExtensionPoint and related code to extension_points package
As extension points will be reused in v3 plugins this makes the code
better separated.
