+++ picard/plugin.py
@@ -33,13 +33,16 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
-
 from collections import defaultdict
 import os.path
 
 from picard import log
-from picard.config import get_config
 from picard.const import USER_PLUGIN_DIR
+from picard.extension_points import (
+    PLUGIN_MODULE_PREFIX,
+    PLUGIN_MODULE_PREFIX_LEN,
+    ExtensionPoint,
+)
 from picard.version import (
     Version,
     VersionError,
@@ -54,62 +57,6 @@ except ImportError:
         if not text:
             return ''
         return text.strip().replace('\n', '<br>\n')
-
-_PLUGIN_MODULE_PREFIX = "picard.plugins."
-_PLUGIN_MODULE_PREFIX_LEN = len(_PLUGIN_MODULE_PREFIX)
-
-_extension_points = []
-
-
-def _unregister_module_extensions(module):
-    for ep in _extension_points:
-        ep.unregister_module(module)
-
-
-class ExtensionPoint:
-
-    def __init__(self, label=None):
-        if label is None:
-            import uuid
-            self.label = uuid.uuid4()
-        else:
-            self.label = label
-        self.__dict = defaultdict(list)
-        _extension_points.append(self)
-
-    def register(self, module, item):
-        if module.startswith(_PLUGIN_MODULE_PREFIX):
-            name = module[_PLUGIN_MODULE_PREFIX_LEN:]
-            log.debug("ExtensionPoint: %s register <- plugin=%r item=%r", self.label, name, item)
-        else:
-            name = None
-            # uncomment to debug internal extensions loaded at startup
-            # print("ExtensionPoint: %s register <- item=%r" % (self.label, item))
-        self.__dict[name].append(item)
-
-    def unregister_module(self, name):
-        try:
-            del self.__dict[name]
-        except KeyError:
-            # NOTE: needed due to defaultdict behaviour:
-            # >>> d = defaultdict(list)
-            # >>> del d['a']
-            # KeyError: 'a'
-            # >>> d['a']
-            # []
-            # >>> del d['a']
-            # >>> #^^ no exception, after first read
-            pass
-
-    def __iter__(self):
-        config = get_config()
-        enabled_plugins = config.setting['enabled_plugins'] if config else []
-        for name in self.__dict:
-            if name is None or name in enabled_plugins:
-                yield from self.__dict[name]
-
-    def __repr__(self):
-        return f"ExtensionPoint(label='{self.label}')"
 
 
 class PluginShared:
@@ -138,8 +85,8 @@ class PluginWrapper(PluginShared):
     @property
     def module_name(self):
         name = self.module.__name__
-        if name.startswith(_PLUGIN_MODULE_PREFIX):
-            name = name[_PLUGIN_MODULE_PREFIX_LEN:]
+        if name.startswith(PLUGIN_MODULE_PREFIX):
+            name = name[PLUGIN_MODULE_PREFIX_LEN:]
         return name
 
     @property
